pyproject.toml
[project]
name = "tulip"
version = "0.1.0"
description = "A Telegram bot written using the python-telegram-bot library"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-telegram-bot>=21.6",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.5.2",
    "sqlalchemy>=2.0.35",
    "alembic>=1.13.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.8",
]
